Must use hash to re-render
Use  a closure in the form of an iife to prevent calling framework logic
directly. Way to change parameters and re-render is restricted to
resetting the url hash. With this new restriction means things will
always work a certain way, and some logic can be removed simplifying the
framework. Likewise the framework is simpler to use; just update
window.location.hash, and the page will re-render to keep in sync.